parallelized computation of distances (#28)
parallelized computation of distances for some of the metrics exist for cdist. Cover the existing euclidean an cosine for knn and mknn.
Significantly accelerates the computation. 
if some other metrics will desired to be used by users (instead the euclidean and cosine for knn and mknn respectively),
then If the user have enough memory and the metric is in distance_metrics, then users can replace the whole loop, that remained for their decision.